Release v3.9.0: CRT-Phosphor radar display, Radar Pulse, quick watchlist add
- Radar Display menu now offers a CRT-Phosphor glow effect and a Radar
  Pulse animation as independent, combinable options alongside the
  existing Green/Amber/Blue color schemes
- Aircraft can now be added to the watchlist directly from the detail
  panel with a single tap ("+"/"-" button next to "QR")

+++ src/i18n.cpp
@@ -23,8 +23,8 @@ const char* t(StringId id) {
     uint8_t lang = SettingsStore::language();
     if (lang >= LANG_COUNT) lang = 0;
 
-    uint8_t idx = (uint8_t)id;
-    if (idx >= (uint8_t)StringId::COUNT) return "?";
+    uint16_t idx = (uint16_t)id;
+    if (idx >= (uint16_t)StringId::COUNT) return "?";
 
     return TABLES[lang][idx];
 }
